More robust handling of swaps (#72)
* fix: more robust handling of tx, int32 -> int64

* test: added test

* test: fix tests

+++ entities/plugin.go
@@ -4,7 +4,7 @@ package entities
 type Plugin interface {
 	// Execute can be called multiple times for the same job
 	// It's plugin responsibility to check if the job is already running and ignore such requests
-	Execute(jobID int32, data []byte, msgCallback MessageCallback) error
+	Execute(jobID int64, data []byte, msgCallback MessageCallback) error
 }
 
 // MessageCallback is callback function which is called by plugin to send message to connector
@@ -12,7 +12,7 @@ type MessageCallback func(PluginMessage) error
 
 // PluginMessage is structure of message sent from plugin to connector
 type PluginMessage struct {
-	JobID      int32
+	JobID      int64
 	Message    string
 	Data       []byte
 	IsError    bool
