+++ plugins/boltz/main_test.go
@@ -74,9 +74,9 @@ func TestExecute(t *testing.T) {
 		BoltzAPI:    bapi.NewBoltzPrivateAPI("https://testapi.boltz.exchange", nil),
 		ChainParams: &chaincfg.TestNet3Params,
 		LnAPI:       mkFakeLndAPI(),
-		jobs:        make(map[int32]interface{}),
+		jobs:        make(map[int64]interface{}),
 		db: &TestDB{data: map[interface{}]interface{}{
-			int32(42): target,
+			int64(42): target,
 		}},
 		Limits: common.SwapLimits{
 			MaxAttempts: 10,
@@ -116,7 +116,7 @@ func TestExecute(t *testing.T) {
 		assert.True(t, ok)
 
 		fmt.Printf("%+v\n", p.db.(*TestDB).data)
-		_, ok = p.db.(*TestDB).data[int32(123)]
+		_, ok = p.db.(*TestDB).data[int64(123)]
 		assert.True(t, ok)
 	})
 
@@ -129,7 +129,7 @@ func TestExecute(t *testing.T) {
 		err := p.Execute(42, target, cs.Callback)
 		require.NoError(t, err)
 
-		_, ok := p.jobs[int32(42)]
+		_, ok := p.jobs[int64(42)]
 		assert.True(t, ok)
 	})
 }
